+++ app/controllers/designs_controller.rb
@@ -155,19 +155,16 @@ class DesignsController < ApplicationController
     @design = @project.designs.find_by_id(params[:design_id]) if @project
     @site = @project.sites.find_by_id(params[:site_id]) if @project
     @emails = params[:emails].to_s.split(/[;\r\n]/).collect{|email| email.strip}.select{|email| not email.blank?}.uniq
-    @date = parse_date(params[:sheet_date])
 
 
     if @project
-      if @design and @site and not @date.blank? and not @emails.blank?
-        # @sheets = @design.batch_sheets!(current_user, @site, @date, @emails)
-        # redirect_to project_sheets_path(@project, sheet_after: @date.strftime("%m/%d/%Y"), sheet_before: @date.strftime("%m/%d/%Y"), site_id: @site.id, design_id: @design.id, user_id: current_user.id), notice: "#{@sheets.count} #{@sheets.count == 1 ? 'sheet was' : 'sheets were'} successfully created."
-        (sheets_created, sheets_ignored) = @design.batch_sheets!(current_user, @site, @date, @emails)
+      if @design and @site and not @emails.blank?
+        (sheets_created, sheets_ignored) = @design.batch_sheets!(current_user, @site, @emails)
         flash[:notice] = "#{sheets_created} #{sheets_created == 1 ? 'sheet was' : 'sheets were'} successfully created." if sheets_created > 0
         flash[:alert] = "#{sheets_ignored} #{sheets_ignored == 1 ? 'sheet was' : 'sheets were'} not created because the #{sheets_ignored == 1 ? 'subject exists' : 'subjects exist'} on a different site." if sheets_ignored > 0
-        redirect_to project_sheets_path(@project, sheet_after: @date.strftime("%m/%d/%Y"), sheet_before: @date.strftime("%m/%d/%Y"), site_id: @site.id, design_id: @design.id, user_id: current_user.id)
+        redirect_to project_sheets_path(@project, site_id: @site.id, design_id: @design.id, user_id: current_user.id)
       else
-        redirect_to batch_project_designs_path(emails: @emails.join('; '), date: @date.blank? ? nil : @date.strftime("%m/%d/%Y"), site_id: @site ? @site.id : nil, design_id: @design ? @design.id : nil ), alert: 'Please select a sheet date, design, site, and valid emails.'
+        redirect_to batch_project_designs_path(emails: @emails.join('; '), site_id: @site ? @site.id : nil, design_id: @design ? @design.id : nil ), alert: 'Please select a design, site, and valid emails.'
       end
     else
       redirect_to root_path
@@ -503,7 +500,7 @@ class DesignsController < ApplicationController
     end
 
     params[:design].slice(
-      :name, :description, :project_id, :option_tokens, :email_template, :email_subject_template, :updater_id, :study_date_name
+      :name, :description, :project_id, :option_tokens, :email_template, :email_subject_template, :updater_id
     )
   end