+++ app/models/design.rb
@@ -30,7 +30,7 @@ class Design < ActiveRecord::Base
 
   # Model Methods
 
-  def batch_sheets!(current_user, site, date, emails)
+  def batch_sheets!(current_user, site, emails)
     new_sheets = []
     ignored_sheets = 0
     emails.each do |email|
@@ -42,7 +42,7 @@ class Design < ActiveRecord::Base
       end
       subject = site.subjects.find_by_email(short_email) unless short_email.blank?
       subject = site.subjects.find_or_create_by_project_id_and_subject_code(site.project_id, email, { user_id: current_user.id, status: 'valid', email: short_email, acrostic: '' }) unless subject
-      sheet = site.project.sheets.find_or_create_by_study_date_and_subject_id_and_design_id(date, subject.id, self.id, { user_id: current_user.id, last_user_id: current_user.id }) if subject
+      sheet = site.project.sheets.find_or_create_by_subject_id_and_design_id(subject.id, self.id, { user_id: current_user.id, last_user_id: current_user.id }) if subject
       sheet.send_external_email!(current_user, short_email) if sheet and not sheet.new_record?
       if sheet and not sheet.new_record?
         new_sheets << sheet