+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Updated Rails to 3.2.12
 
 ### Enhancements
+- Removed `Study Date` as a required field from sheets
+  - All existing sheets have had existing study dates created as a new variable that is attached at the top of the design
 - Updated to Contour 1.2.0.pre3
   - Use bootstrap-datepicker and bootstrap-timepicker provided by Contour
 - Removed references to deprecated `<center>` HTML tag